--- a/aten/src/ATen/Parallel.h
+++ b/aten/src/ATen/Parallel.h
@@ -1,54 +1,59 @@
 #pragma once
 #include <ATen/ATen.h>
 #include <cstddef>
-#include <tbb/tbb.h>
 
+#ifdef _OPENMP
+#include <omp.h>
+#endif
 
 namespace at {
 namespace internal {
-// This needs to be called before the first use of any algorithm such as
-// parallel or it will have no effect and the default task scheduler is
-// created which uses all available cores.
-// See
-// https://www.threadingbuildingblocks.org/docs/help/reference/task_scheduler/task_scheduler_init_cls.html
-// This does not initializes the number of workers in the market (the overall
-// of workers available to a process). It is merely a request to the market
-// for a certain number of workers. If there are multiple threads making
-// a request at the size of the maximum number of threads, they will
-// be allocated a number proportional to the other requests.
-AT_API void init_tbb_num_threads();
 // This parameter is heuristically chosen to determine the minimum number of
 // work that warrants paralellism. For example, when summing an array, it is
 // deemed inefficient to parallelise over arrays shorter than 32768. Further,
 // no parallel algorithm (such as parallel_reduce) should split work into
 // smaller than GRAIN_SIZE chunks.
-constexpr int64_t TBB_GRAIN_SIZE = 32768;
+constexpr int64_t GRAIN_SIZE = 32768;
 } // namespace internal
+
+inline int64_t divup(int64_t x, int64_t y) {
+  return (x + y - 1) / y;
+}
 
 template <class F>
 inline void parallel_for(
-    int64_t begin,
-    int64_t end,
-    int64_t grain_size,
-    const F& f) {
-  internal::init_tbb_num_threads();
-
-#ifdef __PPC64__
-  using default_partitioner_type = tbb::simple_partitioner;
-#else
-  using default_partitioner_type = tbb::affinity_partitioner;
-#endif
-
-  thread_local static default_partitioner_type ap;
-
-  if ((end - begin) < grain_size || get_num_threads() == 1) {
-    f(begin, end);
-  } else {
-    tbb::parallel_for(
-        tbb::blocked_range<int64_t>(begin, end, grain_size),
-        [f](const tbb::blocked_range<int64_t>& r) { f(r.begin(), r.end()); },
-        ap);
+    const int64_t begin,
+    const int64_t end,
+    const int64_t grain_size_,
+    const F f) {
+  const int64_t min_grain_size = divup((end - begin), get_num_threads());
+  const int64_t grain_size = std::max(min_grain_size, grain_size_);
+#pragma omp parallel for if ((end - begin) >= grain_size && get_num_threads() > 1)
+  for (int64_t i = begin; i < end; i += grain_size) {
+    f(i, i + std::min(end - i, grain_size));
   }
+}
+
+template <class scalar_t, class F, class SF>
+inline scalar_t parallel_reduce(
+    const int64_t begin,
+    const int64_t end,
+    const int64_t grain_size_,
+    const scalar_t ident,
+    const F f,
+    const SF sf) {
+  const int64_t min_grain_size = divup((end - begin), get_num_threads());
+  const int64_t grain_size = std::max(min_grain_size, grain_size_);
+  const int64_t num_results = divup((end - begin), grain_size);
+  std::vector<scalar_t> results(num_results);
+  scalar_t* results_data = results.data();
+#pragma omp parallel for if ((end - begin) >= grain_size && get_num_threads() > 1)
+  for (int64_t id = 0; id < num_results; id++) {
+    int64_t i = begin + id * grain_size;
+    results_data[id] = f(i, i + std::min(end - i, grain_size), ident);
+  }
+  return std::accumulate(
+      results_data, results_data + results.size(), ident, sf);
 }
 
 } // namespace at