This is for performance testing only.

EDIT: This has become necessary.